requirement.txt
beautifulsoup4==4.9.1
boto3==1.12.21
bs4==0.0.1
fake-useragent==0.1.11
lxml==4.5.2
numpy==1.19.2
pandas==1.1.1
pymongo==3.10.1
pymysql==0.9.3
python-dotenv==0.12.0
redis==3.4.1
requests==2.24.0
rq==1.3.0
rq-retry==0.3.0
rq-scheduler==0.9.1
Scrapy==2.3.0
selenium==3.141.0
service-identity==18.1.0
retry==0.9.0
